Back_end/requirements.txt
asgiref
certifi
charset-normalizer
contourpy
cycler
Django
django-cors-headers
djangorestframework
filelock
fonttools
fsspec
idna
intel-openmp
Jinja2
kiwisolver
MarkupSafe
matplotlib
mkl
mpmath
networkx
numpy
packaging
pillow
pyparsing
python-dateutil
requests
six
sqlparse
sympy
tbb
torch
torchvision
typing_extensions
tzdata
urllib3